requirements.txt
Pillow
ipython